fix table exports by moving all table components to table to solve problem with macOS its case insensitive file system
