+++ packages/react-admin-core/src/index.ts
@@ -17,12 +17,3 @@ export * from "./Selected";
 export * from "./Selection";
 export * from "./SelectionApi";
 export * from "./SelectionRoute";
-export * from "./Table";
-export * from "./TableDndOrder";
-export * from "./TableFilterFinalForm";
-export * from "./TableLocalChanges";
-export * from "./TableQuery";
-export * from "./TableQueryContext";
-export * from "./useTableQuery";
-export * from "./withDirtyHandlerApi";
-export * from "./withTableQueryContext";
