+++ packages/react-admin-core/src/table/LocalChangesToolbar.tsx
@@ -2,7 +2,7 @@ import { CircularProgress, Toolbar } from "@material-ui/core";
 import Button from "@material-ui/core/Button";
 import SaveIcon from "@material-ui/icons/Save";
 import * as React from "react";
-import { ITableLocalChangesApi } from "../TableLocalChanges";
+import { ITableLocalChangesApi } from "./TableLocalChanges";
 
 interface IProps {
     tableLocalChangesApi: ITableLocalChangesApi;
