+++ packages/react-admin-core/src/table/index.ts
@@ -1,6 +1,19 @@
+// Export folders
+export * from "./pagingStrategy";
+
+// Export files
 export * from "./AddButton";
 export * from "./Column";
 export * from "./DeleteButton";
 export * from "./LocalChangesToolbar";
 export * from "./Pagination";
-export * from "./pagingStrategy";
+export * from "./Table";
+export * from "./TableDndOrder";
+export * from "./TableFilterFinalForm";
+export * from "./TableLocalChanges";
+export * from "./TableQuery";
+export * from "./TableQueryContext";
+
+export * from "./useTableQuery";
+export * from "./withDirtyHandlerApi";
+export * from "./withTableQueryContext";
